Set Tinker's Tool Leveling to 3x XP and new tools start with 0 modifiers
I think this strikes a good balance with leveling not being *too* overpowered.

In the short term it's fairly easy to get the first 2 or 3 modifiers on a tool/weapon. That balances out the fact that you start with 0 modifiers.

In the long term, however, because of the 3x XP requirement it should take significantly longer to get to a place where you feel like you just have an over-abundance of modifiers.